refactor: implement interface on DataSource
