+++ app/src/main/java/com/team22/soundary/core/auth/di/TokenRepositoryModule.kt
@@ -10,7 +10,7 @@ import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
-abstract class TokenRepositoryModule {
+internal abstract class TokenRepositoryModule {
 
     @Binds
     @Singleton
